Commands used in threads are now saved again. Refactor some of the code into plugins.
