+++ src/plugins/snippets.js
@@ -30,7 +30,7 @@ module.exports = bot => {
   });
 
   // Show or add a snippet
-  addInboxServerCommand('snippet', async (msg, args) => {
+  addInboxServerCommand('snippet', async (msg, args, thread) => {
     const trigger = args[0];
     if (! trigger) return
 
@@ -40,42 +40,42 @@ module.exports = bot => {
     if (snippet) {
       if (text) {
         // If the snippet exists and we're trying to create a new one, inform the user the snippet already exists
-        msg.channel.createMessage(`Snippet "${trigger}" already exists! You can edit or delete it with ${config.prefix}edit_snippet and ${config.prefix}delete_snippet respectively.`);
+        utils.postSystemMessageWithFallback(msg.channel, thread, `Snippet "${trigger}" already exists! You can edit or delete it with ${config.prefix}edit_snippet and ${config.prefix}delete_snippet respectively.`);
       } else {
         // If the snippet exists and we're NOT trying to create a new one, show info about the existing snippet
-        msg.channel.createMessage(`\`${config.snippetPrefix}${trigger}\` replies ${snippet.is_anonymous ? 'anonymously ' : ''}with:\n${snippet.body}`);
+        utils.postSystemMessageWithFallback(msg.channel, thread, `\`${config.snippetPrefix}${trigger}\` replies ${snippet.is_anonymous ? 'anonymously ' : ''}with:\n${snippet.body}`);
       }
     } else {
       if (text) {
         // If the snippet doesn't exist and the user wants to create it, create it
         await snippets.add(trigger, text, false);
-        msg.channel.createMessage(`Snippet "${trigger}" created!`);
+        utils.postSystemMessageWithFallback(msg.channel, thread, `Snippet "${trigger}" created!`);
       } else {
         // If the snippet doesn't exist and the user isn't trying to create it, inform them how to create it
-        msg.channel.createMessage(`Snippet "${trigger}" doesn't exist! You can create it with \`${config.prefix}snippet ${trigger} text\``);
+        utils.postSystemMessageWithFallback(msg.channel, thread, `Snippet "${trigger}" doesn't exist! You can create it with \`${config.prefix}snippet ${trigger} text\``);
       }
     }
   });
 
   bot.registerCommandAlias('s', 'snippet');
 
-  addInboxServerCommand('delete_snippet', async (msg, args) => {
+  addInboxServerCommand('delete_snippet', async (msg, args, thread) => {
     const trigger = args[0];
     if (! trigger) return;
 
     const snippet = await snippets.get(trigger);
     if (! snippet) {
-      msg.channel.createMessage(`Snippet "${trigger}" doesn't exist!`);
+      utils.postSystemMessageWithFallback(msg.channel, thread, `Snippet "${trigger}" doesn't exist!`);
       return;
     }
 
     await snippets.del(trigger);
-    msg.channel.createMessage(`Snippet "${trigger}" deleted!`);
+    utils.postSystemMessageWithFallback(msg.channel, thread, `Snippet "${trigger}" deleted!`);
   });
 
   bot.registerCommandAlias('ds', 'delete_snippet');
 
-  addInboxServerCommand('edit_snippet', async (msg, args) => {
+  addInboxServerCommand('edit_snippet', async (msg, args, thread) => {
     const trigger = args[0];
     if (! trigger) return;
 
@@ -84,23 +84,23 @@ module.exports = bot => {
 
     const snippet = await snippets.get(trigger);
     if (! snippet) {
-      msg.channel.createMessage(`Snippet "${trigger}" doesn't exist!`);
+      utils.postSystemMessageWithFallback(msg.channel, thread, `Snippet "${trigger}" doesn't exist!`);
       return;
     }
 
     await snippets.del(trigger);
     await snippets.add(trigger, text, snippet.isAnonymous);
 
-    msg.channel.createMessage(`Snippet "${trigger}" edited!`);
+    utils.postSystemMessageWithFallback(msg.channel, thread, `Snippet "${trigger}" edited!`);
   });
 
   bot.registerCommandAlias('es', 'edit_snippet');
 
-  addInboxServerCommand('snippets', async msg => {
+  addInboxServerCommand('snippets', async (msg, args, thread) => {
     const allSnippets = await snippets.all();
     const triggers = allSnippets.map(s => s.trigger);
     triggers.sort();
 
-    msg.channel.createMessage(`Available snippets (prefix ${config.snippetPrefix}):\n${triggers.join(', ')}`);
+    utils.postSystemMessageWithFallback(msg.channel, thread, `Available snippets (prefix ${config.snippetPrefix}):\n${triggers.join(', ')}`);
   });
 };
